ai_predict/requirements.txt
ultralytics
opencv-python
paho-mqtt

